+++ Cargo.toml
@@ -10,8 +10,8 @@ members = [
   "dna_c_binding",
   "hdk-rust",
   "net",
-  "net_connection",
-  "net_ipc",
+  #"net_connection",
+  #"net_ipc",
   "nodejs_waiter",
   "sodium",
   "hc_dpki",
