+++ conductor_api/src/lib.rs
@@ -95,7 +95,6 @@ extern crate holochain_core;
 extern crate holochain_core_types;
 extern crate holochain_dpki;
 extern crate holochain_net;
-extern crate holochain_net_ipc;
 extern crate holochain_sodium;
 
 extern crate chrono;
