+++ core/Cargo.toml
@@ -34,7 +34,6 @@ holochain_core_types_derive = { path = "../core_types_derive" }
 holochain_core_types = { path = "../core_types" }
 holochain_cas_implementations = { path = "../cas_implementations" }
 holochain_dpki = { path = "../hc_dpki" }
-holochain_net_connection = { path = "../net_connection" }
 holochain_sodium = { path = "../sodium" }
 boolinator = "2.4.0"
 jsonrpc-ws-server = { git = "https://github.com/paritytech/jsonrpc" }
