+++ core/src/network/reducers/init.rs
@@ -4,7 +4,7 @@ use crate::{
     network::{handler::create_handler, state::NetworkState},
 };
 use holochain_net::p2p_network::P2pNetwork;
-use holochain_net_connection::{
+use holochain_net::connection::{
     json_protocol::{JsonProtocol, TrackDnaData},
     net_connection::NetSend,
 };
