+++ core/src/network/reducers/mod.rs
@@ -36,7 +36,7 @@ use crate::{
     },
 };
 use holochain_core_types::{cas::content::Address, error::HolochainError};
-use holochain_net_connection::{
+use holochain_net::connection::{
     json_protocol::{JsonProtocol, MessageData},
     net_connection::NetSend,
 };
