+++ core/src/network/reducers/respond_get.rs
@@ -4,7 +4,7 @@ use crate::{
     network::{actions::ActionResponse, reducers::send, state::NetworkState},
 };
 use holochain_core_types::{entry::EntryWithMeta, error::HolochainError};
-use holochain_net_connection::json_protocol::{FetchEntryData, FetchEntryResultData, JsonProtocol};
+use holochain_net::connection::json_protocol::{FetchEntryData, FetchEntryResultData, JsonProtocol};
 use std::sync::Arc;
 
 /// Send back to network a HandleFetchEntryResult, no matter what.
