+++ core/src/network/reducers/send_direct_message.rs
@@ -4,7 +4,7 @@ use crate::{
     network::{reducers::send, state::NetworkState},
 };
 use holochain_core_types::error::HolochainError;
-use holochain_net_connection::json_protocol::{JsonProtocol, MessageData};
+use holochain_net::connection::json_protocol::{JsonProtocol, MessageData};
 use std::sync::Arc;
 
 fn inner(
