+++ net/Cargo.toml
@@ -2,15 +2,18 @@
 name = "holochain_net"
 version = "0.0.4-alpha"
 authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
+edition = "2018"
 
 [dependencies]
 failure = "0.1.1"
 holochain_core_types = { path = "../core_types" }
 holochain_core_types_derive = { path = "../core_types_derive" }
-holochain_net_connection = { path = "../net_connection" }
-holochain_net_ipc = { path = "../net_ipc" }
 lazy_static = "1.2"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = { version = "1", features = ["preserve_order"] }
-snowflake = "1"
+rmp-serde = "0.13.7"
+serde_bytes = "0.10.4"
+regex = "1"
+zmq = "0.9.0"
+snowflake = "1.2"
