+++ net/src/connection/json_protocol.rs
