+++ net/src/connection/net_relay.rs
@@ -1,5 +1,5 @@
 use super::NetResult;
-use crate::{net_connection::*, protocol::Protocol};
+use super::{net_connection::*, protocol::Protocol};
 
 /// a simple pass-through NetSend instance
 /// this struct can be use to compose one type of NetWorker into another
