+++ net/src/in_memory/mod.rs
@@ -1,0 +1,5 @@
+//! Module for the in-memory network module
+
+pub mod memory_book;
+pub mod memory_server;
+pub mod memory_worker;
