+++ net/src/ipc/spawn.rs
@@ -2,7 +2,7 @@
 //! This process is expected to output some specific messages on its stdout
 //! that we can process to know its launch state
 
-use holochain_net_connection::{net_connection::NetShutdown, NetResult};
+use crate::connection::{net_connection::NetShutdown, NetResult};
 
 use std::{
     collections::HashMap,
