+++ net/src/lib.rs
@@ -1,3 +1,4 @@
+#![feature(fnbox)]
 #![feature(try_from)]
 #![feature(vec_remove_item)]
 #![warn(unused_extern_crates)]
@@ -8,12 +9,10 @@
 
 #[macro_use]
 extern crate failure;
-extern crate holochain_core_types;
-extern crate holochain_net_connection;
-extern crate holochain_net_ipc;
+#[macro_use]
+pub extern crate holochain_core_types_derive;
 #[macro_use]
 extern crate lazy_static;
-extern crate snowflake;
 // macros used in tests
 #[allow(unused_imports)]
 #[macro_use]
@@ -21,14 +20,12 @@ extern crate serde_json;
 #[allow(unused_imports)]
 #[macro_use]
 extern crate serde_derive;
-#[macro_use]
-pub extern crate holochain_core_types_derive;
 
 pub mod error;
 pub mod ipc_net_worker;
-pub mod memory_book;
-pub mod memory_server;
-pub mod memory_worker;
 pub mod p2p_config;
 pub mod p2p_network;
 pub mod tweetlog;
+pub mod ipc;
+pub mod connection;
+pub mod in_memory;
