+++ net/src/p2p_network.rs
@@ -2,7 +2,7 @@
 //! P2pNetwork instances take a json configuration string
 //! and at load-time instantiate the configured "backend"
 
-use holochain_net_connection::{
+use crate::connection::{
     net_connection::{NetHandler, NetSend, NetWorker, NetWorkerFactory},
     net_connection_thread::NetConnectionThread,
     protocol::Protocol,
@@ -10,7 +10,11 @@ use holochain_net_connection::{
 };
 use std::{thread::sleep, time::Duration};
 
-use super::{ipc_net_worker::IpcNetWorker, memory_worker::InMemoryWorker, p2p_config::*};
+use crate::{
+    ipc_net_worker::IpcNetWorker,
+    in_memory::memory_worker::InMemoryWorker,
+    p2p_config::*,
+};
 
 /// Facade handling a p2p module responsable for the network connection
 /// Holds a NetConnectionThread and implements itself the NetSend Trait
