--- net_ipc/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "holochain_net_ipc"
-version = "0.0.4-alpha"
-authors = ["neonphog <neonphog@gmail.com>"]
-
-[dependencies]
-failure = "0.1.1"
-holochain_net_connection = { path = "../net_connection" }
-lazy_static = "1.2"
-serde_derive = "1.0"
-regex = "1"
-zmq = "0.9.0"
-snowflake = "1.2"
-
-[dev-dependencies]
