+++ test_bin/Cargo.toml
@@ -8,7 +8,6 @@ holochain_core = { path = "../core" }
 holochain_conductor_api = { path = "../conductor_api" }
 holochain_cas_implementations = { path="../cas_implementations" }
 holochain_core_types = { path = "../core_types" }
-holochain_net_connection = { path="../net_connection" }
 holochain_net = { path = "../net" }
 serde_json = { version = "1", features = ["preserve_order"] }
 tempfile = "3"
