+++ test_bin/src/basic_workflows.rs
@@ -1,7 +1,7 @@
 use constants::*;
 use holochain_core_types::cas::content::Address;
 use holochain_net::tweetlog::TWEETLOG;
-use holochain_net_connection::{
+use holochain_net::connection::{
     json_protocol::{ConnectData, JsonProtocol, TrackDnaData},
     net_connection::NetSend,
     NetResult,
